requirements.txt
customtkinter>=5.2.2
Pillow>=10.0.0
rembg>=2.0.50